Added builder for main object. Adjusted observer pattern. Moved example code to own file
